Notes: Modal Template is set
Need: to Finish CSS on them.

FeedBack:
Anything Else That needs to be added to variables list that will be used for the calculations

Please create a list of any other needed variables

My goal in the morning  is to have the modals styled, Do we want to keep the side column that would have an interactable form or scratch the idea or modify it to just display the responses figure it'll be like a javascript code where within the form there are <p></p> that have been classed that can be id in java and run an array to input the data into the static form.